Add load method

+++ pocket/dao/dao.go
@@ -6,7 +6,7 @@ type DAO interface {
 	AddUser(string) (int64, error)
 	AddArticle(model.Row) error
 	IsUser(int64) (bool, error)
-	GetCountForDates(int, int) (*model.Stats, error)
+	GetCountForDates(int64, int64) (*model.Stats, error)
 	GetArticle(int64) (*model.Row, error)
 	UpdateArticle(*model.Row) error
 	GetLastAdded() (int64, error)
